Revert "Simplify library API"
This reverts commit 21b398a9503582dbc8a847edcfaa96df82eea43e.

The API I removed in the reverted commit was useful for dynamically
constructing a callback signature to verify. While this has never been
needed for simple use cases, I can envision it being useful in cases
where the signature of the callback is not known at compile time, such
as the proposed command overloading system proposed in
pmmp/PocketMine-MP#6518.